ml-evaluation-worker/pyproject.toml
[tool.poetry]
name = "ml-evaluation-app"
version = "0.1.0"
description = ""
authors = ["Igor Bobrukiewicz <igor.bobrukiewicz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "^2.2.3"
uvicorn = "^0.34.2"
pydantic-settings = "^2.9.1"
spacy = "^3.8.5"
transformers = "^4.51.3"
torch = "^2.7.0"
bs4 = "^0.0.2"
celery = "^5.5.2"
flower = "^2.0.1"
fastapi = "^0.115.12"
redis = "^6.1.0"
pika = "^1.3.2"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9"
pgvector = "^0.4.1"
sentence-transformers = "^4.1.0"
openai = "^1.86.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest = "^8.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"